@@ -53,13 +53,8 @@ module.exports = {
         use: {
           loader: "babel-loader",
           options: {
-            "presets": [
-              "@babel/preset-env", 
-              "@babel/preset-react",
-              {
-                'plugins': ["@babel/plugin-transform-runtime", '@babel/plugin-proposal-class-properties']
-              }
-            ]
+            "presets": ["@babel/preset-env",["@babel/preset-react", { "runtime": "automatic" }]],
+            "plugins": ["@babel/plugin-transform-runtime", "@babel/plugin-proposal-class-properties"]
           }
         }
       },
